feat: add message ignored
